pyproject.toml
[tool.poetry]
name = "ai-conference-organizer"
version = "0.0.1"
description = "ns bounty project #2"
authors = ["Dixing Xu <dixingxu@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
pymupdf = "^1.24.10"
python-multipart = "^0.0.12"
replit = "^4.1.0"
python-dotenv = "^1.0.1"
openai = "^1.51.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
